+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Plugin now have access to much more data from Neucore.
   - Plugins can now add menu items to the "Services" menu.
 - Change: Line breaks in the Base64-encoded authentication tokens of applications are removed.
+- Fix: It was possible to accept a group application even if the player was not allowed to be a member of that group.
 
 ## 1.44.0